refactor(cli): split packages/cli/src/html.ts into packages/cli/src/html/ (#31)
The single 1399-LOC html.ts has grown unmanageable since the #14 report
redesign. Split into five focused modules plus a shared helper:

  packages/cli/src/html/
    index.ts             orchestration + public types re-export (38 LOC)
    render-shell.ts      <html>…<body> envelope + scripts (154 LOC)
    render-dashboard.ts  stat strip, toc, dashboard sections (251 LOC)
    render-patterns.ts   pattern cards + empty state (65 LOC)
    styles.ts            CSS block as a template-string constant (798 LOC)
    types.ts             RecentRun / HotFile / KindBreakdown / RunFailure (43 LOC)
    utils.ts             esc, severity, formatting, color helpers (144 LOC)

`generateHtml`'s signature and output are preserved byte-for-byte;
existing html.test.ts passes unchanged, and a direct diff of the
generated document against HEAD shows zero differences.

Adds per-renderer unit tests:
  - render-patterns.test.ts  — empty state, escaping, severity bucketing,
                               snapshot of a representative pattern
  - render-dashboard.test.ts — summary stats fallback, toc anchoring,
                               kind-breakdown percentages, SHA truncation

No visual or API changes. Pure refactor.
