+++ packages/cli/src/html/render-shell.ts
@@ -1,0 +1,154 @@
+import { STYLES } from './styles'
+
+/** Pre-rendered section strings injected into the report envelope. */
+export interface ShellSections {
+  statStrip: string
+  toc: string
+  patterns: string
+  dashboard: string
+}
+
+/** Wraps the pre-rendered sections in the full `<!DOCTYPE html>` document —
+ *  theme-probe script, inline CSS, layout shell, and the end-of-body
+ *  interactivity script. No inputs are escaped here; callers escape at the
+ *  point where untrusted strings enter the section HTML. */
+export function renderShell(
+  sections: ShellSections,
+  params: { patternCount: number; windowDays: number },
+): string {
+  const { patternCount, windowDays } = params
+  const plural = patternCount === 1 ? 'pattern' : 'patterns'
+
+  return `<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <title>flaky-tests — ${patternCount} ${plural} detected</title>
+  <script>
+    (function () {
+      try {
+        var stored = localStorage.getItem('flaky-theme');
+        if (stored === 'light') document.documentElement.classList.add('light');
+        else if (stored === 'dark') document.documentElement.classList.add('dark');
+      } catch (err) { /* localStorage may be unavailable on file:// */ }
+    })();
+  </script>
+  <style>
+${STYLES}  </style>
+</head>
+<body>
+  <div class="layout">
+    <header class="page-header">
+      <div class="brand">
+        <span class="brand-mark">✗</span>
+        <span>flaky-tests</span>
+        <span class="brand-sub">report</span>
+      </div>
+      <div class="header-right">
+        <div class="header-meta">window: ${windowDays}d &nbsp;·&nbsp; ${patternCount} ${plural}</div>
+        <button id="theme-toggle" class="theme-toggle" type="button" aria-label="Toggle theme">
+          <svg class="icon-sun" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" aria-hidden="true"><circle cx="12" cy="12" r="4"/><path d="M12 2v2M12 20v2M4.93 4.93l1.41 1.41M17.66 17.66l1.41 1.41M2 12h2M20 12h2M4.93 19.07l1.41-1.41M17.66 6.34l1.41-1.41"/></svg>
+          <svg class="icon-moon" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" aria-hidden="true"><path d="M21 12.79A9 9 0 1 1 11.21 3 7 7 0 0 0 21 12.79z"/></svg>
+        </button>
+      </div>
+    </header>
+
+    ${sections.statStrip}
+
+    ${sections.toc}
+
+    <main class="main-col">
+      <h2 class="section-title patterns-title">Detected Patterns</h2>
+      ${sections.patterns}
+      ${sections.dashboard}
+    </main>
+
+    <footer class="footer">
+      Generated by <a href="https://github.com/brewpirate/flaky-tests">flaky-tests</a>
+      &nbsp;·&nbsp; click Copy on any card or use the disclosure to select-all manually
+    </footer>
+  </div>
+  <!--
+    Opt-in progressive enhancement: the report is fully usable without JS.
+    This single inline block wires the Copy buttons; nothing else depends on it.
+  -->
+  <script>
+    (function () {
+      var toggle = document.getElementById('theme-toggle');
+      if (toggle) {
+        toggle.addEventListener('click', function () {
+          var root = document.documentElement;
+          var isLight = root.classList.contains('light');
+          if (!isLight) {
+            var prefersLight = window.matchMedia && window.matchMedia('(prefers-color-scheme: light)').matches;
+            var hasDark = root.classList.contains('dark');
+            isLight = !hasDark && prefersLight;
+          }
+          if (isLight) {
+            root.classList.remove('light');
+            root.classList.add('dark');
+            try { localStorage.setItem('flaky-theme', 'dark'); } catch (err) { /* ignore */ }
+          } else {
+            root.classList.remove('dark');
+            root.classList.add('light');
+            try { localStorage.setItem('flaky-theme', 'light'); } catch (err) { /* ignore */ }
+          }
+        });
+      }
+      var buttons = document.querySelectorAll('.copy-btn[data-prompt-id]');
+      function flash(btn) {
+        btn.classList.add('copied');
+        btn.textContent = 'Copied ✓';
+        setTimeout(function () {
+          btn.classList.remove('copied');
+          btn.textContent = btn.getAttribute('data-original') || 'Copy';
+        }, 1500);
+      }
+      function fallbackSelect(node) {
+        try {
+          var range = document.createRange();
+          range.selectNodeContents(node);
+          var sel = window.getSelection();
+          if (sel) { sel.removeAllRanges(); sel.addRange(range); }
+        } catch (err) { /* best-effort fallback */ }
+      }
+      buttons.forEach(function (btn) {
+        btn.addEventListener('click', function (event) {
+          event.preventDefault();
+          event.stopPropagation();
+          var id = btn.getAttribute('data-prompt-id');
+          var target = id ? document.getElementById(id) : null;
+          if (!target) return;
+          var text = target.textContent || '';
+          try {
+            if (navigator.clipboard && navigator.clipboard.writeText) {
+              navigator.clipboard.writeText(text).then(function () { flash(btn); }, function () { fallbackSelect(target); });
+            } else { fallbackSelect(target); }
+          } catch (err) { fallbackSelect(target); }
+        });
+      });
+      var runRows = document.querySelectorAll('.run-row[data-run-id]');
+      function toggleRun(row) {
+        var id = row.getAttribute('data-run-id');
+        if (!id) return;
+        var sibling = document.querySelector('.run-failures[data-run-id="' + id.replace(/"/g, '\\"') + '"]');
+        if (!sibling) return;
+        var isHidden = sibling.hasAttribute('hidden');
+        if (isHidden) { sibling.removeAttribute('hidden'); row.classList.add('expanded'); }
+        else { sibling.setAttribute('hidden', ''); row.classList.remove('expanded'); }
+      }
+      runRows.forEach(function (row) {
+        row.addEventListener('click', function (event) {
+          if (event.target && event.target.closest && event.target.closest('a')) return;
+          toggleRun(row);
+        });
+        row.addEventListener('keydown', function (event) {
+          if (event.key === 'Enter' || event.key === ' ') { event.preventDefault(); toggleRun(row); }
+        });
+      });
+    })();
+  </script>
+</body>
+</html>`
+}
